// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/KnowledgeLink.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface KnowledgeLink
 */
export interface KnowledgeLink {
    /**
     * 
     * @type {string}
     * @memberof KnowledgeLink
     */
    knowledgeId: string;
    /**
     * 
     * @type {boolean}
     * @memberof KnowledgeLink
     */
    primary: boolean;
    /**
     * 
     * @type {number}
     * @memberof KnowledgeLink
     */
    confidence?: number | null;
}

/**
 * Check if a given object implements the KnowledgeLink interface.
 */
export function instanceOfKnowledgeLink(value: object): value is KnowledgeLink {
    if (!('knowledgeId' in value) || value['knowledgeId'] === undefined) return false;
    if (!('primary' in value) || value['primary'] === undefined) return false;
    return true;
}

export function KnowledgeLinkFromJSON(json: any): KnowledgeLink {
    return KnowledgeLinkFromJSONTyped(json, false);
}

export function KnowledgeLinkFromJSONTyped(json: any, ignoreDiscriminator: boolean): KnowledgeLink {
    if (json == null) {
        return json;
    }
    return {
        
        'knowledgeId': json['knowledgeId'],
        'primary': json['primary'],
        'confidence': json['confidence'] == null ? undefined : json['confidence'],
    };
}

export function KnowledgeLinkToJSON(json: any): KnowledgeLink {
    return KnowledgeLinkToJSONTyped(json, false);
}

export function KnowledgeLinkToJSONTyped(value?: KnowledgeLink | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'knowledgeId': value['knowledgeId'],
        'primary': value['primary'],
        'confidence': value['confidence'],
    };
}