// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/WrongCreate.ts
export interface WrongCreate {
    /**
     * 
     * @type {string}
     * @memberof WrongCreate
     */
    studentId: string;
    /**
     * 
     * @type {string}
     * @memberof WrongCreate
     */
    subjectId: string;
    /**
     * 
     * @type {WrongSource}
     * @memberof WrongCreate
     */
    sourceType: WrongSource;
    /**
     * 
     * @type {string}
     * @memberof WrongCreate
     */
    questionText: string;
    /**
     * 
     * @type {string}
     * @memberof WrongCreate
     */
    studentAnswer?: string | null;
    /**
     * 
     * @type {string}
     * @memberof WrongCreate
     */
    correctAnswer?: string | null;
    /**
     * 
     * @type {string}
     * @memberof WrongCreate
     */
    analysisText?: string | null;
    /**
     * 
     * @type {string}
     * @memberof WrongCreate
     */
    errorType?: string | null;
    /**
     * 
     * @type {number}
     * @memberof WrongCreate
     */
    difficulty?: number | null;
    /**
     * 
     * @type {Array<KnowledgeLink>}
     * @memberof WrongCreate
     */
    knowledgePoints?: Array<KnowledgeLink>;
}